refactor: create ProductsModule
